+++ TinyNvidiaUpdateChecker/Handlers/GPUHandler.cs
@@ -1,4 +1,3 @@
-using Microsoft.Win32;
 using Newtonsoft.Json;
 using System;
 using System.Collections.Generic;