+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/) 
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [Unreleased]
+### Fixed
+- Proper Quadro / RTX Enterprise driver identification (does not apply to experimental metadata)
+
 ## [1.27.0] - 2026-09-15
 ### Added
 - Support for multiple drivers, prompt UI now shows all drivers available for the detected GPU, and allows you to choose which one to install